[Release 2.10.5-ex120] Feat: Forge 标签 & 导线, GUI 重做, 导线连接渲染
Auto generated by using ./gradlew runData

说是重做 GUI, 其实就是改了阴影、调了坐标

受影响: 所有机器内文本阴影

+++ src/main/java/ic2/core/Ic2Potion.java
@@ -19,10 +19,10 @@ public class Ic2Potion extends MobEffect
 		if (this == radiation)
 		{
 			if (Ic2DamageSource.radiation == null)
-		{
-			Ic2DamageSource.init(entity.level().registryAccess());
-		}
-		entity.hurt(Ic2DamageSource.radiation, amplifier / 100 + 0.5F);
+			{
+				Ic2DamageSource.init(entity.level().registryAccess());
+			}
+			entity.hurt(Ic2DamageSource.radiation, (float) amplifier / 100 + 0.5F);
 		}
 	}
